fix: lint, format, and test errors

+++ caldera/consensus/block_rewards.py
@@ -15,7 +15,7 @@ def calculate_pool_reward(height: uint32) -> uint64:
     """
 
     if height == 0:
-        return 0
+        return uint64(0)
     elif height < 3 * _blocks_per_year:
         return uint64(int((7 / 8) * 10 * _mojo_per_caldera))
     elif height < 6 * _blocks_per_year:
@@ -38,7 +38,7 @@ def calculate_base_farmer_reward(height: uint32) -> uint64:
     rates increase continuously.
     """
     if height == 0:
-        return 0
+        return uint64(0)
     elif height < 3 * _blocks_per_year:
         return uint64(int((1 / 8) * 10 * _mojo_per_caldera))
     elif height < 6 * _blocks_per_year:
@@ -61,7 +61,7 @@ def calculate_postfarm_reward(height: uint32) -> uint64:
     rates increase continuously.
     """
     if height == 0:
-        return 0
+        return uint64(0)
     elif height < 3 * _blocks_per_year:
         return uint64(int((1 / 100) * 10 * _mojo_per_caldera))
     elif height < 6 * _blocks_per_year:
